+++ tests/test_sync.py
@@ -4,16 +4,21 @@
 
 import datetime as dt
 
+from unittest.mock import Mock
+
 import pytest
 import tenacity
 
 import stamina
 
-from stamina._core import _make_stop
+from stamina._core import _make_stop, _StopBefore
 
 
 @pytest.mark.parametrize("attempts", [None, 1])
-@pytest.mark.parametrize("timeout", [None, 1, dt.timedelta(days=1)])
+@pytest.mark.parametrize(
+    "timeout",
+    [None, 1, dt.timedelta(days=1), dt.datetime.now(tz=dt.timezone.utc)],
+)
 @pytest.mark.parametrize("duration", [1, dt.timedelta(days=1)])
 def test_ok(attempts, timeout, duration):
     """
@@ -34,8 +39,17 @@ def test_ok(attempts, timeout, duration):
     assert 42 == f()
 
 
+@pytest.mark.parametrize(
+    "timeout",
+    [
+        None,
+        1,
+        dt.timedelta(days=1),
+        dt.datetime.now(tz=dt.timezone.utc) + dt.timedelta(days=1),
+    ],
+)
 @pytest.mark.parametrize("duration", [0, dt.timedelta(days=0)])
-def test_retries(duration):
+def test_retries(duration, timeout):
     """
     Retries if the specific error is raised.
     """
@@ -43,13 +57,14 @@ def test_retries(duration):
 
     @stamina.retry(
         on=ValueError,
+        timeout=timeout,
         wait_max=duration,
         wait_initial=duration,
         wait_jitter=duration,
     )
     def f():
         nonlocal i
-        if i == 0:
+        if i < 1:
             i += 1
             raise ValueError
 
@@ -113,3 +128,27 @@ class TestMakeStop:
         If all conditions are None, return stop_never.
         """
         assert tenacity.stop_never is _make_stop(attempts=None, timeout=None)
+
+
+class TestStopBefore:
+    @pytest.mark.parametrize(
+        "deadline, wait_initial, wait_exp_base, wait_jitter",
+        [
+            (dt.datetime.now(tz=dt.timezone.utc), 1, 1, 1),
+            (dt.datetime.now().astimezone(), 1, 1, 1),  # noqa: DTZ005
+        ],
+    )
+    def test_deadline_expired(
+        self, deadline, wait_initial, wait_exp_base, wait_jitter
+    ):
+        """
+        If deadline could possibly be overstepped, return True to stop.
+        """
+        sb = _StopBefore(
+            deadline,
+            wait_initial=wait_initial,
+            wait_exp_base=wait_exp_base,
+            wait_jitter=wait_jitter,
+        )
+
+        assert sb(Mock(attempt_number=1))